+++ test/models/thing_test.rb
@@ -245,4 +245,46 @@ class ThingTest < ActiveSupport::TestCase
     assert thing.links.exists?(link_type: :wiki)
     assert_not thing.links.exists?(link_type: :asset)
   end
+
+  test "normalizes a MAC address to lowercase colon form" do
+    thing = Thing.create!(name: "Switch", mac_address: "94-2A-6F-26-C6-CB")
+
+    assert_equal "94:2a:6f:26:c6:cb", thing.mac_address
+  end
+
+  test "accepts a blank MAC address" do
+    thing = Thing.new(name: "No MAC", mac_address: "  ")
+
+    assert thing.valid?
+    assert_nil thing.mac_address
+  end
+
+  test "rejects a MAC address that is not six octets" do
+    thing = Thing.new(name: "Bad MAC", mac_address: "94:2a:6f")
+
+    assert_not thing.valid?
+    assert_includes thing.errors[:mac_address], "must be a valid MAC address"
+  end
+
+  test "requires a unique MAC address" do
+    thing = Thing.new(name: "Duplicate MAC", mac_address: things(:router).mac_address)
+
+    assert_not thing.valid?
+    assert_includes thing.errors[:mac_address], "has already been taken"
+  end
+
+  test "finds things by MAC address" do
+    assert_includes Thing.search("94:2a:6f"), things(:router)
+  end
+
+  test "deleting a thing stops its UniFi devices from recreating it" do
+    thing = things(:router)
+    device = unifi_devices(:rack_switch)
+
+    thing.destroy!
+
+    device.reload
+    assert device.ignored?
+    assert_nil device.thing_id
+  end
 end